Some performane improvements+long for vertex count in vertexbuffer

+++ Graphics/Model/Mesh.cs
@@ -20,7 +20,7 @@
         public override void Draw()
         {
             GraphicsDevice.VertexBuffer = VB;
-            GraphicsDevice.DrawPrimitives(PrimitiveType.Triangles, 0, VB.VertexCount);
+            GraphicsDevice.DrawPrimitives(PrimitiveType.Triangles, 0, (int)VB.VertexCount);
         }
     }
 }
